requirements.txt
# Microsoft 365 Agents SDK for Python
# Requires Python 3.9+ (3.12 recommended)

# Core dependencies
python-dotenv
aiohttp

# Microsoft Agents SDK packages
microsoft-agents-activity
microsoft-agents-hosting-core
microsoft-agents-hosting-aiohttp
microsoft-agents-authentication-msal

# Teams channel support
microsoft-agents-hosting-teams

# Copilot Studio integration
microsoft-agents-copilotstudio-client

# Model Context Protocol (MCP) for Agent A
mcp>=1.0.0

# Microsoft Graph API for M365 access
msgraph-sdk>=1.0.0

# MSAL for authentication
msal>=1.26.0

# Storage extensions (optional - uncomment if needed)
# microsoft-agents-storage-blob
# microsoft-agents-storage-cosmos
